towards a gcc 3.0.3 compileable gift and bugfix for perl-disabled configuration

+++ libMRML/include/CIDRelevanceLevelPair.h
@@ -129,9 +129,9 @@ public:
   inline bool operator()(const CIDRelevanceLevelPair& l,
 			 const CIDRelevanceLevelPair& t){
     return 
-      fabs(l.getID())
+      fabs(double(l.getID()))
       <
-      fabs(t.getID());
+      fabs(double(t.getID()));
   };
 };
